refactor: consolidate integration tests and fix LocalBranchExists
- Remove check_integration_test.go (tests already in remove_integration_test.go)
- Fix LocalBranchExists to use ExitError type check instead of exit code 1
  (git rev-parse returns exit code 128 for non-existent refs)
- Update init_integration_test.go to use t.Context()
- Remove duplicate CheckSkipsDirtySubmodule subtest from remove tests
